Разбивает форму отзыва на отдельные компоненты
