+++ project/src/components/pages/add-review/add-review.jsx
@@ -1,19 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import {PosterModifier, FilmInfo} from '../../../const';
-import {getStarsValue} from '../../../utils/film-util';
+import {PosterModifier} from '../../../const';
 
-import RatingStar from '../../ui/rating-star/rating-star';
 import Preview from '../../ui/film-card/preview/preview';
 import Poster from '../../ui/film-card/poster/poster';
-
-const getRatingStar = (value) => (
-  <RatingStar
-    key={value}
-    count={value}
-  />
-);
+import ReviewForm from '../../ui/review-form/review-form';
 
 function AddReview({id, name, poster, backgroundImage}) {
   return (
@@ -34,29 +26,7 @@ function AddReview({id, name, poster, backgroundImage}) {
       </div>
 
       <div className="add-review">
-        <form action="#" className="add-review__form">
-          <div className="rating">
-            <div className="rating__stars">
-              {getStarsValue(FilmInfo.MAX_STAR_COUNT).map(getRatingStar)}
-            </div>
-          </div>
-
-          <div className="add-review__text">
-            <textarea
-              className="add-review__textarea"
-              name="review-text"
-              id="review-text"
-              placeholder="Review text"
-              defaultValue=""
-            />
-
-            <div className="add-review__submit">
-              <button className="add-review__btn" type="submit">
-                Post
-              </button>
-            </div>
-          </div>
-        </form>
+        <ReviewForm />
       </div>
     </section>
   );
