+++ project/src/components/ui/review-form/review-submit-button/review-submit-button.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+function ReviewSubmitButton() {
+  return (
+    <button className="add-review__btn" type="submit">
+      Post
+    </button>
+  );
+}
+
+export default ReviewSubmitButton;
